requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
google-generativeai==0.8.3
langchain==0.3.1
langchain-community==0.3.1
pypdf==4.3.1
docx2txt==0.8
python-telegram-bot==21.5
slack-bolt==1.20.1
streamlit==1.38.0
httpx==0.27.2
python-multipart==0.0.12
numpy==1.26.4
